requirements.txt
Flask
Pillow
ultralytics
